New function load_jupytext_configuration_file
